[IMP] l10n_dk,l10n_dk_nemhandel_response: merge response module
Merges the module l10n_dk_nemhandel_response (created in stable to
support Application Responses) into l10n_dk.
