+++ .weblate.json
@@ -52,11 +52,6 @@
                 "new_base": "addons/account_peppol/i18n/account_peppol.pot"
             },
             {
-                "name": "account_peppol_response",
-                "filemask": "addons/account_peppol_response/i18n/*.po",
-                "new_base": "addons/account_peppol_response/i18n/account_peppol_response.pot"
-            },
-            {
                 "name": "account_qr_code_emv",
                 "filemask": "addons/account_qr_code_emv/i18n/*.po",
                 "new_base": "addons/account_qr_code_emv/i18n/account_qr_code_emv.pot"
@@ -2154,12 +2149,6 @@
                 "name": "l10n_dk_fik",
                 "filemask": "addons/l10n_dk_fik/i18n/*.po",
                 "new_base": "addons/l10n_dk_fik/i18n/l10n_dk_fik.pot",
-                "language_regex": "^(da)$"
-            },
-            {
-                "name": "l10n_dk_nemhandel_response",
-                "filemask": "addons/l10n_dk_nemhandel_response/i18n/*.po",
-                "new_base": "addons/l10n_dk_nemhandel_response/i18n/l10n_dk_nemhandel_response.pot",
                 "language_regex": "^(da)$"
             },
             {
