+++ addons/l10n_dk/models/__init__.py
@@ -5,6 +5,7 @@ from . import account_edi_xml_oioubl_21
 from . import account_journal
 from . import account_move
 from . import account_move_send
+from . import nemhandel_response
 from . import res_company
 from . import res_config_settings
 from . import res_partner
