+++ addons/l10n_dk/models/account_move.py
@@ -12,6 +12,8 @@ class AccountMove(models.Model):
             ('to_send', 'Queued'),
             ('processing', 'Pending Reception'),
             ('done', 'Done'),
+            ('BusinessAccept', 'Approved'),
+            ('BusinessReject', 'Rejected'),
             ('error', 'Error'),
         ],
         compute='_compute_nemhandel_move_state',
@@ -19,11 +21,15 @@ class AccountMove(models.Model):
         string='Nemhandel status',
         copy=False,
     )
+    nemhandel_response_ids = fields.One2many('nemhandel.response', 'move_id')
+    nemhandel_can_send_response = fields.Boolean(compute='_compute_nemhandel_can_send_response')
 
-    @api.depends('state')
+    @api.depends('state', 'nemhandel_response_ids.nemhandel_state')
     def _compute_nemhandel_move_state(self):
         for move in self:
-            if all([
+            if valid_status := move.nemhandel_response_ids.filtered(lambda r: r.nemhandel_state == 'done').mapped('response_code'):
+                move.nemhandel_move_state = valid_status[0]
+            elif all([
                 move.company_id.l10n_dk_nemhandel_proxy_state == 'receiver',
                 move.commercial_partner_id.nemhandel_verification_state == 'valid',
                 move.state == 'posted',
@@ -39,6 +45,18 @@ class AccountMove(models.Model):
                 move.nemhandel_move_state = False
             else:
                 move.nemhandel_move_state = move.nemhandel_move_state
+
+    @api.depends("nemhandel_response_ids.nemhandel_state")
+    def _compute_nemhandel_can_send_response(self):
+        for move in self:
+            move.nemhandel_can_send_response = (
+                move.nemhandel_message_uuid
+                and move.move_type in ('in_invoice', 'in_refund')
+                and not move.nemhandel_response_ids.filtered(
+                    lambda r: r.nemhandel_state == 'not_serviced' or r.nemhandel_state != 'error',
+                )
+                and move.partner_id.nemhandel_response_support
+            )
 
     @api.model
     def _get_ubl_cii_builder_from_xml_tree(self, tree):
@@ -61,6 +79,11 @@ class AccountMove(models.Model):
 
         return super()._get_import_file_type(file_data)
 
+    def _post(self, soft=True):
+        res = super()._post(soft)
+        self.action_nemhandel_send_approval_response()
+        return res
+
     def action_cancel_nemhandel_documents(self):
         # if the nemhandel_move_state is processing/done
         # then it means it has been already sent to nemhandel proxy and we can't cancel
@@ -79,3 +102,37 @@ class AccountMove(models.Model):
         if ubl_cii_format == 'oioubl_21' and (not self.partner_id.vat or self.partner_id._get_nemhandel_verification_state(ubl_cii_format) != 'valid'):
             return False
         return res
+
+    def button_cancel(self):
+        res = super().button_cancel()
+        if action := self.action_nemhandel_open_rejection_wizard():
+            action['context'] = {'cancel_res': res}
+            return action
+        return res
+
+    def action_nemhandel_send_approval_response(self):
+        moves_to_respond_by_company = self.filtered('nemhandel_can_send_response').grouped('company_id')
+        for company in moves_to_respond_by_company:
+            company.nemhandel_edi_user._nemhandel_send_response(moves_to_respond_by_company[company], 'BusinessAccept')
+
+    def action_nemhandel_open_rejection_wizard(self):
+        nemhandel_moves = self.filtered('nemhandel_can_send_response')
+        if nemhandel_moves:
+            return {
+                'type': 'ir.actions.act_window',
+                'name': self.env._("Reject Nemhandel Document"),
+                'view_mode': 'form',
+                'res_model': 'nemhandel.rejection.wizard',
+                'target': 'new',
+                'res_id': self.env['nemhandel.rejection.wizard'].create({'move_ids': nemhandel_moves.ids}).id,
+            }
+        return {}
+
+    def action_open_nemhandel_reponses(self):
+        return {
+            'type': 'ir.actions.act_window',
+            'name': self.env._("Nemhandel Responses"),
+            'view_mode': 'list',
+            'res_model': 'nemhandel.response',
+            'domain': [('id', 'in', self.nemhandel_response_ids.ids)],
+        }
