+++ addons/l10n_dk/security/ir.model.access.csv
@@ -1,2 +1,4 @@
 id,name,model_id:id,group_id:id,perm_read,perm_write,perm_create,perm_unlink
 l10n_dk_registration,l10n.dk.registration,model_nemhandel_registration,account.group_account_invoice,1,1,1,1
+nemhandel_rejection_wizard,nemhandel.rejection.wizard,model_nemhandel_rejection_wizard,account.group_account_invoice,1,1,1,1
+nemhandel_response,nemhandel.response,model_nemhandel_response,account.group_account_invoice,1,1,1,1
