+++ addons/l10n_dk/views/nemhandel_response_views.xml
