+++ addons/l10n_dk/wizard/__init__.py
@@ -1,3 +1,4 @@
 from . import account_move_send_batch_wizard
 from . import account_move_send_wizard
 from . import nemhandel_registration
+from . import nemhandel_rejection_wizard
