--- addons/l10n_dk_nemhandel_response/__init__.py
@@ -1,2 +1,0 @@
-from . import models
-from . import wizard
