--- addons/l10n_dk_nemhandel_response/models/__init__.py
@@ -1,4 +1,0 @@
-from . import account_edi_proxy_user
-from . import account_move
-from . import nemhandel_response
-from . import res_partner
