--- addons/l10n_dk_nemhandel_response/security/ir.model.access.csv
@@ -1,3 +1,0 @@
-id,name,model_id:id,group_id:id,perm_read,perm_write,perm_create,perm_unlink
-nemhandel_rejection_wizard,nemhandel.rejection.wizard,model_nemhandel_rejection_wizard,account.group_account_invoice,1,1,1,1
-nemhandel_response,nemhandel.response,model_nemhandel_response,account.group_account_invoice,1,1,1,1
