--- addons/l10n_dk_nemhandel_response/wizard/__init__.py
@@ -1,1 +1,0 @@
-from . import nemhandel_rejection_wizard
